requirements.txt
fastapi
uvicorn[standard]
python-multipart
supabase
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pytesseract
Pillow
PyPDF2
langchain
langchain-community
faiss-cpu
ollama
pydantic